added about section
