--- surf-spot frontend/src/__tests__/Home.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Home from '../Home';
-import { mount } from 'enzyme'
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<Home />, div);
-});
