+++ surf-spot-frontend/public/favicon.ico
