+++ surf-spot-frontend/public/manifest.json
