+++ surf-spot-frontend/src/Beach.js
